+++ src/client/search.rs
@@ -235,7 +235,7 @@ pub fn search_documents(
     let mut response: Response = request
         .send()
         .map_err(|e| e.context(ErrorKind::HttpRequestFailed))?
-        .general_err_handler(StatusCode::OK)?;
+        .general_err_handler(&[StatusCode::OK])?;
     debug!("Response: '{:#?}'", response);
 
     let result = response.json().map_err(|e| {